assistant/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pinecone-client==3.2.2
prisma==0.11.0
python-dotenv==1.0.0
openai==1.14.0
httpx==0.27.2
langchain-openai==0.0.7
langchain-pinecone==0.0.3
langchain-core==0.1.26
langchain-community==0.0.20
langchain==0.2.17
langsmith==0.1.17
aiohttp==3.9.0
python-multipart==0.0.6
twilio==8.10.0